+++ src/Core/Tests/CoreTests.cpp
@@ -1,7 +1,7 @@
 /**
  * @file CoreTests.cpp
  * @copyright This code is licensed under the 3-clause BSD license.\n
- *            Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.\n
+ *            Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.\n
  *            See LICENSE.txt for details.
  */
 #include "DummyInterface.h"
@@ -50,6 +50,39 @@ TEST(CoreModules, ModuleCorrectness) {
 TEST(CoreModules, StringSplit) {
   const std::vector<std::string> expected{{"a", "b", "c", "d"}};
   ASSERT_EQ(Detail::split("a;b;c;d", ';'), expected);
+}
+
+class FakeModule : public Scine::Core::Module {
+ public:
+  std::string name() const noexcept final {
+    return "FakeModule";
+  }
+
+  boost::any get(const std::string& /* interface */, const std::string& /* model */) const final {
+    return NULL;
+  }
+
+  bool has(const std::string& /* interface */, const std::string& /* model */) const noexcept final {
+    return false;
+  }
+
+  std::vector<std::string> announceInterfaces() const noexcept final {
+    return {};
+  }
+
+  std::vector<std::string> announceModels(const std::string& /* interface */) const noexcept final {
+    return {};
+  }
+
+  static std::shared_ptr<Scine::Core::Module> make() {
+    return std::make_shared<FakeModule>();
+  }
+};
+
+TEST(CoreModules, DirectModuleLoad) {
+  auto& manager = ModuleManager::getInstance();
+  auto module = std::make_shared<FakeModule>();
+  manager.load(module);
 }
 
 TEST(ObjectWithLog, Basics) {